requirements.txt
fastapi
uvicorn[standard]
biopython
pandas
numpy
scikit-learn
nltk
spacy
gensim
matplotlib
seaborn
python-dotenv
googletrans==3.1.0a0
aiohttp
openpyxl